Reorganize codebase into v2 namespace and update README
Major code reorganization:

**Namespace Restructure:**
- Move all modern "neo" code to v2 namespace (src/params_proto/v2/)
- Update main __init__.py to export from v2
- Keep v1 namespace for backward compatibility
- Move all v2 tests to tests/test_v2/ directory

**README Updates:**
- Add type annotations to all example code
- Update installation instructions (neotype)
- Add inline comments for clarity
- Update imports to use new `from params_proto import proto` pattern
- Document v3.0.0 release with uv migration notes

**Code Quality:**
- Consistent type hints across examples
- Better code organization with clear version separation
- Updated flag syntax (bool type annotations instead of Flag())

All tests passing after reorganization.
