--- src/params_proto/__init__.py
@@ -1,4 +1,0 @@
-from .proto import *
-from . import hyper
-from . import utils
-from .utils import read_deps
